Absorbing the old files into new cli_ files, updated tests and docs
